==> analysis/08-calculate-diversity.Rmd <==
---
title: "Social diversity"
author: "Qingqing Chen"
date: "Last compiled date: `r format(Sys.time(), '%d %B, %Y')`"
output: github_document
editor_options:
  chunk_output_type: console
---

<style>
body {text-align: justify}
</style>


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width = 10, fig.height = 10, warning = FALSE, message = FALSE)
library(tidyverse)
library(dplyr)
library(sf)
library(tmap)
library(RColorBrewer)
library(vegan) ##calculate diversity
library(purrrogress)
library(here)
source(here("R/viz.R"))
```

To measure diversity, we use Shannon's diversity index (H) as it is commonly used for species diversity and accounts for both richness and evenness of the species present. The higher the Shannon's index (H) value is, the more diverse the origins of visitors are for inflow diversity. Conversely, for outflow diversity a higher H means that locals visit a more diverse set of neighbourhoods throughout the city.


## Load necessary data 

```{r}
# aggregated grids 
grids <- read_sf(here("data/derived_data/spatial_hex_grid.shp"), quiet = T) %>% 
  st_transform(crs = 3414)

# Singapore planning areas 
sg_subzone <- read_sf(here("data/raw_data/sg-subzone/MP14_SUBZONE_NO_SEA_PL.shp")) %>%
  st_transform(crs = 3414) %>% 
  st_make_valid() %>% 
  group_by(PLN_AREA_N) %>% 
  dplyr::summarise()

# grid sectors 
grid_sectors <- readRDS(here("data/derived_data/grid_sectors.rds"))

# tweets
df_tweets <- readRDS(here("data/derived_data/deidentified_sg_tweets_updated.rds")) 

# qualified grid cells 
considered_grids <-  df_tweets %>% 
  group_by(grid_id) %>% 
  dplyr::summarise(n_tweets = n(), n_users = n_distinct(u_id)) %>% 
  filter(n_tweets >= 100 & n_users >= 5)

# usrs in each grid cell 
users_in_grids <- df_tweets %>% distinct(u_id, grid_id)

# identified home location of users 
identified_hms <- read_csv(here("data/derived_data/identified_hms.csv"))

# distinct users in grids: local or visitor 
distinct_users <- readRDS(here("data/derived_data/distinct_users.rds"))
```


## Diversity analysis with Shannon's index 

```{r}
# function used for diversity analysis
cal_diversity <- function(grid_index, data, grid_sectors, grids, type = "inflow"){
  # get sectors of input grid 
  sectors <- grid_sectors %>% filter(grid_id == grid_index) 
  # get sector areas intersect with SG boundary 
  sectors_corssSG <- st_intersection(sectors, sg_subzone) %>% 
    dplyr::select(-PLN_AREA_N) %>% 
    unique() %>%
    group_by(sector_id, grid_id) %>% 
    dplyr::summarise() %>% 
    ungroup()
  
  # assign users to sectors 
  if(type == "inflow"){
    # assign visitors' home locations to sectors 
    users_in_sectors <- st_join(data, sectors_corssSG, largest = TRUE) %>% 
      distinct(u_id, home, .keep_all = TRUE) %>% 
      dplyr::select(sector_id, grid_id, u_id, home) %>% 
      st_set_geometry(NULL) %>% 
      left_join(., sectors_corssSG) %>% # change the geometry to sectors' geometry
      st_sf()  # convert to sf object 
  }else if(type == "outflow"){
    # assign locals visited locations to sectors 
    users_in_sectors <- st_join(data, sectors_corssSG, largest = T) %>% 
      distinct(u_id, out_to_grid, .keep_all = TRUE) %>% 
      dplyr::select(sector_id, grid_id, u_id, out_to_grid) %>% 
      st_set_geometry(NULL) %>% 
      left_join(., sectors_corssSG) %>% # change the geometry to sectors' geometry
      st_sf()
  }
  
  # calculate Shannon's index  
  output <- users_in_sectors %>%
      group_by(sector_id) %>%
      dplyr::summarise(n_user = n_distinct(u_id)) %>%
      ungroup() %>%
      mutate(area_km_square = as.numeric(st_area(.)/1000000)) %>%
      mutate(user_density = n_user/area_km_square) %>%
      st_set_geometry(NULL) %>%
      distinct(sector_id, user_density) %>%
      spread(sector_id, user_density) %>%
      diversity(index = "shannon") 
  return(output)
}
```

### Inflow diversity 

The inflow diversity is saved under `data/derived_data/` directory. 
```{r}
# Get visitors in each grid 
visitors_in_grid <-  distinct_users %>% 
  filter(type == "visitor") %>% 
  left_join(., grids, by = c("home" = "grid_id")) %>% 
  st_as_sf(crs = 3414) %>% 
  nest(data = c(u_id, home, type, geometry))
  
# visited grid 
visitors_in_grid[1, ]

# visitors with home locations 
visitors_in_grid[1, ]$data[[1]]
```


```{r}
# calculate diversity 
if(file.exists(here("data/derived_data/inflow_diversity.rds"))){
  inflow_diversity <- readRDS(here("data/derived_data/inflow_diversity.rds"))
}else{
  inflow_diversity <- visitors_in_grid %>% 
    filter(grid_id %in% considered_grids$grid_id) %>% 
    mutate(div_shannon = map2_dbl(grid_id, data, with_progress(function(x, y) cal_diversity(x, y, grid_sectors, grids, type = "inflow")))) %>% 
    dplyr::select(-data) 
  
  inflow_diversity <- inflow_diversity %>% 
    mutate(norm_div_shannon = (div_shannon - min(div_shannon))/(max(div_shannon) - min(div_shannon))) %>% 
    left_join(., grids, by = c("grid_id" = "grid_id")) %>% 
    st_sf(crs = 3414)
  saveRDS(inflow_diversity, file = here("data/derived_data/inflow_diversity.rds"))
}
head(inflow_diversity)
```

### Outflow diversity

The grids except home that each local visited are saved under `data/derived_data/` directory. 

```{r}
# Get other grids except home that locals visited 
# locals in home grids
locals_in_grid <- distinct_users %>% 
  filter(type == "local") %>% 
  nest(data = c(u_id, home, type))

# grids outside home grid that each local visited 
get_locals_out_grids <- function(grid_id, data, users_in_grids, grids){
  # locals in home grid 
  locals <-  data$u_id %>% unique()
  # home grid id 
  locals_home <- grid_id
  output <- users_in_grids %>% 
    filter(u_id %in% locals) %>%  ## get grids that locals sent tweets 
    filter(!grid_id %in% locals_home) %>% ## remove grid that is the same as the home
    left_join(., grids) %>% ## get out to grid geometry
    rename(out_to_grid = grid_id) %>% 
    st_sf() 
  return(output)
}

if(file.exists(here("data/derived_data/locals_out_grids.rds"))){
  locals_out_grids <- readRDS(here("data/derived_data/locals_out_grids.rds")) 
}else{
  locals_out_grids <- locals_in_grid %>% 
    filter(grid_id %in% considered_grids$grid_id) %>% 
    mutate(data = map2(grid_id, data, with_progress(function(x, y) get_locals_out_grids(x, y, users_in_grids, grids))))
  saveRDS(locals_out_grids, file = here("data/derived_data/locals_out_grids.rds"))
}

# home grid 
locals_out_grids[1, ]

# locals in home grid with their visited grids outside home 
locals_out_grids[1, ]$data[[1]]
```

The outflow diversity is saved under `data/derived_data/` directory. 
```{r}
# calculate outflow diversity 
if(file.exists(here("data/derived_data/outflow_diversity.rds"))){
  outflow_diversity <- readRDS(here("data/derived_data/outflow_diversity.rds"))
}else{
  outflow_diversity <- locals_out_grids %>% 
    filter(grid_id %in% considered_grids$grid_id) %>% 
    mutate(div_shannon = map2_dbl(grid_id, data, with_progress(function(x, y) cal_diversity(x, y, grid_sectors, grids, type = "outflow")))) %>% 
    dplyr::select(-data) 
  
  outflow_diversity <- outflow_diversity %>% 
    mutate(norm_div_shannon = (div_shannon - min(div_shannon))/(max(div_shannon) - min(div_shannon)), 
           norm_div_shannon = round(norm_div_shannon, 2)) %>% 
    left_join(., grids, by = c("grid_id" = "grid_id")) %>% 
    st_sf(crs = 3414)
    
  saveRDS(outflow_diversity, file = here("data/derived_data/outflow_diversity.rds"))
}

head(outflow_diversity)
```


## Diversity spatial distribution 

### Inflow diversity 

```{r}
spatial_viz(inflow_diversity, fill_var = "norm_div_shannon", 
            legend_title = "Norm.diversity", 
            main_title = "(a) Diversity of Flows: normalized inflow diversity", 
            main.title_size = 1.4,
            legend.hist_height = 0.25, legend.hist_width = 0.5,
            legend_width = 0.4, legend.hist_size = 0.5,
            legend.title_size = 1.2, legend.text_size = 0.65)
```


```{r fig.height=4}
violin_viz(inflow_diversity, var = "norm_div_shannon",  
           labs.x = "", labs.y = "Norm.diversity", breaks = seq(0, 1, 0.1),
           y.shift = 0.17, x.shift = 1.52, text.size = 4)
```

### Outflow diversity

```{r}
spatial_viz(outflow_diversity, fill_var = "norm_div_shannon", 
            legend_title = "Norm.diversity", 
            main_title = "(b) Diversity of Flows: normalized outflow diversity", 
            main.title_size = 1.4,
            legend.hist_height = 0.25, legend.hist_width = 0.5,
            legend_width = 0.4, legend.hist_size = 0.5,
            legend.title_size = 1.2, legend.text_size = 0.65,
            palette = "Purples")
```


```{r fig.height=4}
violin_viz(outflow_diversity, var = "norm_div_shannon",
           labs.x = "", labs.y = "Norm.diversity", breaks = seq(0, 1, 0.1), 
           y.shift = 0.17, x.shift = 1.52, text.size = 4)
```
